requirements.txt
aiohappyeyeballs==2.4.4 ; python_version >= "3.11" and python_version < "4.0"
aiohttp==3.11.11 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.2 ; python_version >= "3.11" and python_version < "4.0"
attrs==25.1.0 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
frozenlist==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.1.0 ; python_version >= "3.11" and python_version < "4.0"
propcache==0.2.1 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
tenacity==9.0.0 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.18.3 ; python_version >= "3.11" and python_version < "4.0"
